pyproject.toml
[project]
name = "visdiff-impl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.59.8",
    "pillow>=11.1.0",
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.48.0",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "ruff>=0.9.2",
]
